+++ GameTracking/Sql/Procedures/Aggregating Procedures/GameTrack.GetDevelopersAndAverageScore.sql
@@ -1,6 +1,6 @@
 ﻿CREATE OR ALTER PROCEDURE GameTrack.GetDevelopersAndAverageScore
 AS
-SELECT D.DeveloperID, D.[Name], D.Email, D.FoundedDate, D.[Location], D.TeamCount, CAST(AVG(R.Score) AS FLOAT) AS AverageReviewScore
+SELECT D.DeveloperID, D.[Name], D.Email, D.FoundedDate, D.[Location], D.TeamCount, CAST(AVG(R.Score) AS DECIMAL(10, 2)) AS AverageReviewScore
 FROM GameTrack.Game G
 	INNER JOIN GameTrack.GameDeveloper GD ON G.GameID = GD.GameID
 	INNER JOIN GameTrack.Developer D ON GD.DeveloperID = D.DeveloperID
