+++ GameTracking/SqlDeveloperRepository.cs
@@ -161,7 +161,7 @@ namespace GameTracking
 			while (reader.Read())
 			{
 				developers.Add(new Developer(reader.GetInt32(developerIDOrdinal), reader.GetString(nameOrdinal), reader.GetString(emailOrdinal),
-				reader.GetDateTime(foundedDateOrdinal), reader.GetString(locationOrdinal), reader.GetInt32(teamCountOrdinal)) { AverageReviewScore = reader.GetDouble(averageReviewScoreOrdinal)});
+				reader.GetDateTime(foundedDateOrdinal), reader.GetString(locationOrdinal), reader.GetInt32(teamCountOrdinal)) { AverageReviewScore = reader.GetDecimal(averageReviewScoreOrdinal)});
 			}
 
 			return developers;
